pyproject.toml
[project]
name = "readantalk-automation"
version = "0.1.0"
description = ""
authors = [
    {name = "Jay Kim",email = "csjay@snu.ac.kr"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright (>=1.55.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "google-genai (>=1.46.0,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "black (>=25.9.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
